+++ src/crawler/clien/index.js
@@ -75,7 +75,7 @@ class Clien extends CommunityCrawler {
                 this.sortUrl.value,
         );
 
-        const posts = await this.page.evaluate(baseUrl => {
+        return await this.page.evaluate(baseUrl => {
             const lists = document.querySelectorAll('.list_item');
 
             return Array.from(lists)
@@ -116,17 +116,10 @@ class Clien extends CommunityCrawler {
                 })
                 .filter(post => post);
         }, baseUrl);
-
-        return posts.map(post => ({
-            ...post,
-            hasRead: this.postsRead.has(this.title + post.id),
-        }));
     }
 
-    async getPostDetail({ link, id, category }) {
+    async getPostDetail({ title, author, link, id, category }) {
         await this.page.goto(link);
-
-        this.postsRead.add(this.title + id); // set post that you read
 
         const postDetail = await this.page.evaluate(() => {
             const category = document.querySelector(