+++ src/crawler/dcinside/index.js
@@ -60,7 +60,7 @@ class Dcinside extends CommunityCrawler {
             });
         });
 
-        const posts = await this.page.evaluate(() => {
+        return await this.page.evaluate(() => {
             const lists = document.querySelectorAll(
                 '.gall-detail-lst li .gall-detail-lnktb',
             );
@@ -111,17 +111,10 @@ class Dcinside extends CommunityCrawler {
                 })
                 .filter(post => post);
         });
-
-        return posts.map(post => ({
-            ...post,
-            hasRead: this.postsRead.has(this.title + post.id),
-        }));
     }
 
     async getPostDetail({ link, id, category }) {
         await this.page.goto(link);
-
-        this.postsRead.add(this.title + id); // set post that you read
 
         const postDetail = await this.page.evaluate(() => {
             const _title = document.querySelector('.gallview-tit-box .tit');